Feature/SSOR-56-Transaction-Sort (#4)
* SSOR-170 Feature: Pagination implemented as offset. Limit implemented as limit.

* SSOR-170: DTO implemented + Tests for Pagination 100%

* SSOR-170: H2 Memory DB implemented

* SSOR-170: TransactionType implemented

* SSOR-57 Feature: Added TransactionOptions, TransactionType in tests

* SSOR-57: Implemented Filtering, 100% Tests

* SSOR-57 feature: update from remote

* SSOR-56 Feature: Added sortby option

* SSOR-56 Refactor: Adjusted repository return type from list to page

* SSOR-56 Feature: Added sort direction option

* SSOR-56 Tests: Added tests for the modular sort direction

+++ boss-transaction-service/src/main/resources/schema.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS transaction_type (
 
 CREATE TABLE IF NOT EXISTS transaction (
     id INT UNSIGNED NOT NULL AUTO_INCREMENT,
-    type_id INT UNSIGNED NOT NULL,
+    type_id TINYINT UNSIGNED NOT NULL,
     account_id INT UNSIGNED NOT NULL,
     overdraft_id INT UNSIGNED NULL,
     atm_transaction_id INT UNSIGNED NULL,
